Convert Output to jsx
