+++ .eslintignore
@@ -1,3 +1,4 @@
 /lib/*
 /dist/*
+/dist-before/*
 /node_modules/*
